+++ src/main/java/com/moa/user/presentation/HealthCheckController.java
@@ -1,12 +1,14 @@
 package com.moa.user.presentation;
 
+
+import com.moa.global.vo.ApiResult;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import com.moa.user.common.ApiResult;
+
 
 @RestController
 @RequestMapping("/api/v1/user")
@@ -14,9 +16,9 @@ import com.moa.user.common.ApiResult;
 @RequiredArgsConstructor
 public class HealthCheckController {
 
-    @GetMapping("/test")
-    public ResponseEntity<ApiResult<String>> test() {
-        return ResponseEntity.ok(ApiResult.ofSuccess("user test"));
-    }
+	@GetMapping("/test")
+	public ResponseEntity<ApiResult<String>> test() {
+		return ResponseEntity.ok(ApiResult.ofSuccess("user test"));
+	}
 
 }
